Update: Schedule -> ScheduleList 로 변경 및 ChatRoomHeader 컴포넌트 리팩토링 #17
